mod minishell_loop
